LayoutTests/http/tests/security/contentSecurityPolicy/script-redirect-blocked.html

<!DOCTYPE html>
<html>
<head>
<script src="resources/dump-as-text.js"></script>
<script src="resources/wait-until-done.js"></script>
<meta http-equiv="Content-Security-Policy" content="script-src http://127.0.0.1:8000/resources/redirect.py 'unsafe-inline'">
<script src="http://127.0.0.1:8000/resources/redirect.py?code=307&url=http%3A//127.0.0.1%3A8000/resources/redirect.py%3Furl=http%3A//localhost%3A8000/resources/alert-fail.js"></script>
<script>
    // Expect the blocked URI to be the requested origin, not the redirect target.
    document.addEventListener('securitypolicyviolation', e => {
        document.body.innerHTML += `blockedURI = <b>${e.blockedURI}</b><br/><br/>`;
        window.testRunner.notifyDone();
    });
</script>
</head>
<body>
</body>
</html>
