LayoutTests/http/tests/security/contentSecurityPolicy/script-redirect-blocked-expected.txt

CONSOLE MESSAGE: Refused to load the script 'http://localhost:8000/security/contentSecurityPolicy/resources/alert-fail.js' because it violates the following Content Security Policy directive: "script-src http://127.0.0.1:8000/resources/redirect.php 'unsafe-inline'".

ALERT: PASS
